+++ packages/Core/CHANGELOG.md
@@ -4,6 +4,12 @@
 - Use [unfetch](https://github.com/developit/unfetch), [sockette](https://github.com/lukeed/sockette) and [Websocket Node](https://github.com/theturtle32/WebSocket-Node) for new client/server code.
 - Add SDL schema support. [Link](https://www.prisma.io/blog/graphql-sdl-schema-definition-language-6755bcb9ce51/).
 
+## [2.1.3] 1-30-2019
+- Added feature detection for WebSQL in safari to fix issue [115](https://github.com/ClickSimply/Nano-SQL/issues/115).
+- Added new filter "configTableSystem" to resolve issue [117](https://github.com/ClickSimply/Nano-SQL/issues/117).
+- Fixed issue with new query listener feature.
+- Used `getAll` api in IndexedDB to dramatically increase performance of select statements in browsers that support the api.  Resolves issue [118](https://github.com/ClickSimply/Nano-SQL/issues/118).
+
 ## [2.1.2] 1-30-2019
 - **BREAKING CHANGE** Removed `.streamEvent` function, you can now just pass a boolean into the `.stream` function to enable events: `.stream(onRow, complete, error, useEvents)`.
 - `objectsEqual` utility function is now an alias of the `fast-deep-equal` library, should have better performance for row comparisons and index updates.